Aula04/multithread_process/pyproject.toml
[project]
name = "multithread-process"
version = "0.1.0"
description = ""
authors = [
    {name = "Leandro Oliveira",email = "leandro.gooli@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "rembg (>=2.0.65,<3.0.0)"
]

[tool.poetry]
packages = [{include = "multithread_process", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
